dashboard_helmet/requirements.txt
# Python Runtime: 3.10.19

PySide6==6.10.0
pyserial==3.5
requests==2.32.5